clinic_backend/requirements.txt
# Django Core
Django>=4.2,<5.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS Headers
django-cors-headers>=4.3.0

# Google Authentication
google-auth>=2.25.0

# Database (optional - for production)
# psycopg2-binary>=2.9.9

# Environment variables (recommended)
python-dotenv>=1.0.0
